requirements-dev.txt
pyarrow>=20.0.0
datasets>=3.0.0
fastapi>=0.120.1
fastapi_utils==0.8.0
loguru==0.7.3
mangum==0.19.0
pre-commit==4.5.0
pydantic>=2.5.0,<3.0.0
pytest-xdist==3.8.0
pytest-asyncio==1.3.0
# ASGI Server - Version constrained by litellm[proxy]==1.81.3 in rest-api
uvicorn>=0.31.1,<0.32.0

# AWS SDK - Version constrained by litellm[proxy]==1.81.3 in rest-api
boto3==1.40.76
tiktoken==0.12.0
python-docx==1.2.0
pypdf==6.4.1
langchain==1.2.7
langchain-community==0.4.1
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
cachetools==6.2.2
--only-binary=pyarrow,lxml,psycopg2-binary

# Testing
pytest==9.0.2
pytest-cov==7.0.0
moto[proxy]==5.1.18
coverage>=7.10.6
lxml==6.0.2
opensearch-py==3.1.0
requests_aws4auth==1.3.1
PyJWT==2.10.1
psycopg2-binary==2.9.11
cachetools==6.2.2

# MCP Workbench testing
fastmcp>=2.0.0

# Development
black==25.12.0
flake8==7.3.0
mypy==1.19.0
poetry==2.2.1
yq==3.4.3
